Bug fixed #4 - and update #1
Ok! Now the main function does not require to specify if scaling is performed or not. Based on the type of sequence used (in this project defined by t2Ssfp[123]), the scaling is selected internally.

Also, with the updated functionality of the bids_layout.py file, the bash code is not hard-coded for a maximum of two runs but instead any number of them. Besides, this updated functionality helped to reduce the coding length.